rename [...rest] to module
